+++ modulo4/projeto-toDoList/request.rest
@@ -1,0 +1,1 @@
+GET http://localhost:3003/user
