+++ modulo4/projeto-toDoList/src/index.ts
@@ -1,27 +1,14 @@
 import express, { Express } from 'express'
 import cors from 'cors'
 import { AddressInfo } from "net"
-import knex from "knex"
-import dotenv from "dotenv"
-
-dotenv.config()
-
-export const connection = knex({
-    client: "mysql",
-    connection: {
-        host: process.env.DB_HOST,
-        port: 3306,
-        user: process.env.DB_USER,
-        password: process.env.DB_PASS,
-        database: process.env.DB_NAME
-    }
-})
+import getUser from './endpoints/getUser'
 
 const app: Express = express()
 
 app.use(express.json())
 app.use(cors())
 
+app.get('/user', getUser)
 
 const server = app.listen(process.env.PORT || 3003, () => {
     if (server) {
